Use fp32 in iota kernel example

+++ oneapi-rs/examples/kernel_launch_derive.rs
@@ -22,21 +22,21 @@ namespace syclexp = sycl::ext::oneapi::experimental;
 
 extern "C"
 SYCL_EXT_ONEAPI_FUNCTION_PROPERTY((syclexp::nd_range_kernel<1>))
-void iota(double start, double *ptr) {
+void iota(float start, float *ptr) {
     size_t id = syclext::this_work_item::get_nd_item<1>().get_global_linear_id();
-    ptr[id] = start + static_cast<double>(id);
+    ptr[id] = start + static_cast<float>(id);
 }
 "#;
 
 #[derive(KernelArgumentList)]
 struct IotaArgs<'a> {
-    start: f64,
-    ptr: &'a mut Buffer<f64, UsmAllocator<SharedAllocator>>,
+    start: f32,
+    ptr: &'a mut Buffer<f32, UsmAllocator<SharedAllocator>>,
 }
 
 fn main() {
     let mut queue = Queue::new();
-    let mut buffer = queue.alloc_shared::<f64>(1024).wait();
+    let mut buffer = queue.alloc_shared::<f32>(1024).wait();
 
     let kernel = queue
         .get_context()
@@ -49,7 +49,7 @@ fn main() {
             NdRange::new([1024], [16]),
             &kernel,
             IotaArgs {
-                start: 3.14,
+                start: 3.14_f32,
                 ptr: &mut buffer,
             },
         )
